Fix: make price in budgets and accounts integer.

+++ database/migrations/2024_07_20_153945_create_budgets_table.php
@@ -14,7 +14,7 @@ return new class () extends Migration {
     {
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
-            $table->float('size', 2);
+            $table->integer('size');
             $table->unsignedBigInteger('category_id');
             $table->unsignedBigInteger('user_id');
